ui/lims/server: fix long requests being closed
Summary:
Requests such as downloading the reports or kicking-off a large number
of analysis runs (e.g. source conta reruns) will lead to the LIMS proxy
server abruptly closing the connection after 2 minutes. The LIMS API
server still successfully performs the request, but the result doesn't
get returned due to the proxy closing the connection to nginx.

The arbitrarily 2 minutes stems from node's built-in `http` module that
sets the timeout. This module is used by basically every server and
proxy implementation that we use (e.g. `express`, `graphql-yoga`, and
`http-proxy-middleware` which uses `http-proxy`). In order to make sure
the request doesn't get closed by any one of these modules, we need to
increase the timeout for all of them. This sets the timeouts for these
modules to 24h to match what is set for LIMS's nginx proxy_read_timeout.

The timeouts are set through 2 new command-line arguments (defaulting
to 24h):
- `timeout`
- `proxyTimeout`

`timeout` is used for specifying the timeouts for "incoming" requests (i.e.
from client to proxy/server). On the other hand, `proxyTimeout` is used
for specifying the timeouts for "outgoing" requests (i.e. from proxy/server
to target and back to client).

Ref {P205}
Fix T23413

Test Plan:
1. Run local LIMS UI server to start the proxy.
2. Create a LIMS API server that returns a result after 3 minutes.
3. `curl` the LIMS API server through the proxy (i.e. `localhost:9000`).
4. Verify that the response gets returned.

Reviewers: ecarrel, jsingh, nsawas, jzhao, O12 LIMS UI

Reviewed By: jsingh, O12 LIMS UI

Subscribers: amich, cji, arinehouse

Maniphest Tasks: T23413

Differential Revision: https://phabricator.grailbio.com/D32632

fbshipit-source-id: 9bc7a8d